+++ src/WordPress/ShortcodeRegistrar.php
@@ -17,6 +17,7 @@ final class ShortcodeRegistrar
     {
         $map = [
             'opentt_matches_grid' => 'shortcode_matches_grid',
+            'opentt_featured_match' => 'shortcode_featured_match',
             'opentt_standings_table' => 'shortcode_standings_table',
             'opentt_match_games' => 'shortcode_games_list',
             'opentt_h2h' => 'shortcode_h2h',